refactor(report): share spec config matching across coverage and mock usage
