+++ core/src/main/kotlin/io/specmatic/core/report/SpecEndpoint.kt
@@ -1,0 +1,10 @@
+package io.specmatic.core.report
+
+import io.specmatic.reporter.ctrf.model.CtrfSpecConfig
+import io.specmatic.reporter.model.OpenAPIOperation
+
+interface SpecEndpoint {
+    val specification: String?
+    fun toOpenApiOperation(): OpenAPIOperation
+    fun toCtrfSpecConfig(): CtrfSpecConfig
+}
